Moves FeatureDevSessionContext to core module to avoid deduping in toolkit and for codescan utility.
